Refactor: Authentication, plugin and package manager
- Extract Basic Authentication from Settings, and add in its own Settings class and Request Handler. Includes interfaces / abstract classes to support the addition of other authentication types.
- Extract Plugin class from SatisPress class, which then acts as the composition root. Bootstrap this in the main plugin file.
- Extract Packages-related methods into PackageManager class, out of SatisPress class.
- Remove singleton pattern from SatisPress class.

SatisPress is still a bit of a God class.
