+++ src/Composable.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * Composable interface
+ *
+ * @package SatisPress
+ * @license GPL-2.0-or-later
+ * @since 0.3.0
+ */
+
+namespace SatisPress;
+
+/**
+ * Segregated interface of something that should be composed.
+ */
+interface Composable {
+	public function compose();
+}
