+++ src/SatisPress.php
@@ -10,42 +10,25 @@
 namespace SatisPress;
 
 /**
- * Main plugin class.
+ * SatisPress class.
  *
  * @since 0.1.0
  */
 class SatisPress {
 	/**
-	 * The main SatisPress instance.
+	 * Package manager.
 	 *
-	 * @since 0.1.0
-	 * @var SatisPress
+	 * @var string
 	 */
-	private static $instance;
+	protected $package_manager;
 
 	/**
-	 * Main plugin instance.
+	 * Initialise SatisPress object.
 	 *
-	 * @since 0.1.0
-	 *
-	 * @return SatisPress
+	 * @param PackageManager $package_manager Package manager.
 	 */
-	public static function instance() {
-		if ( null === self::$instance ) {
-			self::$instance = new self();
-		}
-
-		return self::$instance;
-	}
-
-	/**
-	 * Constructor.
-	 *
-	 * @since 0.1.0
-	 * @see SatisPress::instance();
-	 */
-	private function __construct() {
-		add_action( 'plugins_loaded', [ $this, 'load' ] );
+	public function __construct( PackageManager $package_manager ) {
+		$this->package_manager = $package_manager;
 	}
 
 	/**
@@ -54,23 +37,9 @@ class SatisPress {
 	 * @since 0.2.0
 	 */
 	public function load() {
-		$htaccess_handler = new Htaccess( $this->cache_path() );
-
 		if ( is_admin() ) {
-			$manage_screen = new Admin\Plugins();
-			$manage_screen->load();
-
-			$settings_screen = new Admin\Settings( $htaccess_handler );
-			$settings_screen->load();
-
 			add_action( 'admin_init', [ $this, 'register_assets' ] );
 		}
-
-		$basic_auth = new Authentication\Basic( $htaccess_handler );
-		$basic_auth->load();
-
-		$limit_login_attempts = new Integration\LimitLoginAttempts();
-		$limit_login_attempts->load();
 
 		add_action( 'init', [ $this, 'add_rewrite_rules' ] );
 		add_filter( 'query_vars', [ $this, 'query_vars' ] );
@@ -85,8 +54,6 @@ class SatisPress {
 		// Delete the 'satispress_packages_json' transient.
 		add_action( 'upgrader_process_complete', [ $this, 'flush_packages_json_cache' ] );
 		add_action( 'set_site_transient_update_plugins', [ $this, 'flush_packages_json_cache' ] );
-
-		register_activation_hook( __FILE__, [ $this, 'activate' ] );
 	}
 
 	/**
@@ -158,7 +125,7 @@ class SatisPress {
 
 		if ( ! $json ) {
 			$data     = [];
-			$packages = $this->get_packages();
+			$packages = $this->package_manager->get_packages();
 
 			foreach ( $packages as $slug => $package ) {
 				$data[ $package->get_package_name() ] = $package->get_package_definition();
@@ -180,101 +147,6 @@ class SatisPress {
 	}
 
 	/**
-	 * Retrieve a package by its slug and type.
-	 *
-	 * @since 0.2.0
-	 *
-	 * @param string $slug Package slug (plugin basename or theme directory name).
-	 * @param string $type Package type.
-	 *
-	 * @return Package
-	 */
-	public function get_package( $slug, $type ) {
-		$package_factory = new PackageFactory();
-
-		return $package_factory->create( $type, $slug, $this->cache_path() );
-	}
-
-	/**
-	 * Retrieve a list of packages.
-	 *
-	 * @since 0.2.0
-	 *
-	 * @return array
-	 */
-	public function get_packages() {
-		$packages  = [];
-		$whitelist = $this->get_whitelist();
-
-		foreach ( $whitelist as $type => $identifiers ) {
-			if ( empty( $identifiers ) ) {
-				continue;
-			}
-
-			foreach ( $identifiers as $identifier ) {
-				$package = $this->get_package( $identifier, $type );
-				if ( $package && $package->is_installed() && '' !== $package->get_version_normalized() ) {
-					$packages[ $package->get_slug() ] = $package;
-				}
-			}
-		}
-
-		return $packages;
-	}
-
-	/**
-	 * Retrieve a list of whitelisted packages.
-	 *
-	 * Plugins should be added to the whitelist by hooking into the
-	 * 'satispress_plugins' filter and appending a plugin's basename to the
-	 * array. The basename is the main plugin file's relative path from the
-	 * root plugin directory. Example: simple-image-widget/simple-image-widget.php
-	 *
-	 * Themes should be added by hooking into the 'satispress_themes' filter and
-	 * appending the name of the theme directory. Example: genesis
-	 *
-	 * @since 0.2.0
-	 *
-	 * @return string
-	 */
-	protected function get_whitelist() {
-		$plugins = apply_filters( 'satispress_plugins', [] );
-		$themes  = apply_filters( 'satispress_themes', [] );
-
-		// @todo Implement these through a filter instead.
-		$options = (array) get_option( 'satispress_plugins' );
-		$plugins = array_filter( array_unique( array_merge( $plugins, $options ) ) );
-
-		$options = (array) get_option( 'satispress_themes', [] );
-		$themes  = array_filter( array_unique( array_merge( $themes, $options ) ) );
-
-		return [
-			'plugin' => $plugins,
-			'theme'  => $themes,
-		];
-	}
-
-	/**
-	 * Retrieve the path where packages are cached.
-	 *
-	 * Defaults to 'wp-content/uploads/satispress/'.
-	 *
-	 * @since 0.1.0
-	 *
-	 * @return string
-	 */
-	public function cache_path() {
-		$uploads = wp_upload_dir();
-		$path    = trailingslashit( $uploads['basedir'] ) . 'satispress/';
-
-		if ( ! file_exists( $path ) ) {
-			wp_mkdir_p( $path );
-		}
-
-		return apply_filters( 'satispress_cache_path', $path );
-	}
-
-	/**
 	 * Cache the current version of a plugin before it's udpated.
 	 *
 	 * @since 0.2.0
@@ -291,7 +163,7 @@ class SatisPress {
 		$type = ( isset( $data['plugin'] ) ) ? 'plugin' : 'theme';
 		$slug = $data[ $type ];
 
-		$package = $this->get_package( $data[ $type ], $type );
+		$package = $this->package_manager->get_package( $data[ $type ], $type );
 		if ( $package ) {
 			$package->archive();
 		}
@@ -387,16 +259,5 @@ class SatisPress {
 	 */
 	public function flush_packages_json_cache() {
 		delete_transient( 'satispress_packages_json' );
-	}
-
-	/**
-	 * Functionality during activation.
-	 *
-	 * Sets a flag to flush rewrite rules on the request after activation.
-	 *
-	 * @since 0.1.0
-	 */
-	public function activate() {
-		update_option( 'satispress_flush_rewrite_rules', 'yes' );
 	}
 }
